add sha256 for `tmux-pomodoro-plus`
It builds now, and `tmux.conf` looks right, but I'm not seeing it in the status line.

It also seems like the default keybindings have some collisions. For example, `<tmux-prefix> p` usually goes to the `p`revious window, but this plugin uses it "to toggle between starting/pausing a Pomodoro/break".